requirements.txt
tornado
toro
pika
redis
mockredispy
futures
